last commit

+++ resources/views/data.blade.php
@@ -2,7 +2,7 @@
 @section("content")
 <div class="container">
     <div class="row">
-        <h2 class="text-center">Available test data entries</h2>
+        <h2 class="text-center">Pieejamie testa scenāriji</h2>
     </div>
 
     <div class="text-center">
@@ -15,16 +15,19 @@
                     <div class="list-group-item">
                         <h3>{{ $data_entry->name }}</h3>
                         <a href="/data/{{ $data_entry->id }}/download/slots" class="waves-effect waves-light btn blue">
-                            <button type="button" class="btn btn-success"> Download Slots </button>
+                            <button type="button" class="btn btn-success"> Lejupielādēt reklāmas pauzes</button>
                         </a>
                         <a href="/data/{{ $data_entry->id }}/download/ads" class="waves-effect waves-light btn blue">
-                            <button type="button" class="btn btn-success"> Download Ads </button>
+                            <button type="button" class="btn btn-success"> Lejupielādēt reklāmas</button>
                         </a>
-                        <a href="/data/{{ $data_entry->id }}/delete" class="waves-effect waves-light btn red">
-                            <button type="button" class="btn btn-danger"> Delete </button>
-                        </a>
+                        @if(Auth::User()->role == '2')
+                            <a href="/data/{{ $data_entry->id }}/delete" class="waves-effect waves-light btn red">
+    {{--                            <button type="button" class="btn btn-danger"> Delete </button>--}}
+                                <img src="/trash.png"/>
+                            </a>
+                        @endif
                         <a href="/data/{{ $data_entry->id }}/showalgorithms" class="waves-effect waves-light btn green">
-                            <button type="button" class="btn btn-primary"> Already tested algorithms </button>
+                            <button type="button" class="btn btn-primary"> Testētie algoritmi</button>
                         </a>
                     </div>
                 @endforeach
